# openapi.yaml
...
openapi: 3.1.0
info:
  title: Silicon Hook API
  version: 0.5.0
  summary: Signed webhook endpoints with ordered, acknowledged delivery to Silicons.
  description: |
    Silicon Hook gives each Silicon signed webhook endpoints at
    `https://hook.teamofsilicons.com/silicon/{silicon_id}/{endpoint_key}`.
    Every request to an active endpoint is answered with `webhook.ok`. Requests
    that verify against the hook's signature policy are retained for 14 days
    and delivered in order over the WebSocket stream at `/api/v1/ws` or by
    polling `/api/v1/silicons/{silicon_id}/deliveries`; requests that do not
    verify are withheld and kept in a separate 14-day blocked log. Timestamps
    are UTC RFC 3339 values. Endpoint keys route requests but never
    authenticate them. IAM test app_secret selectors choose isolated sandboxes via
    X-Hook-Test-App-Secret; they never grant actor authority. Deprecated API v1
    contracts sunset only after seven request-free days. Contract accounting is
    independent of optional diagnostic telemetry.
  contact:
    email: engineering@teamofsilicons.com
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
- url: https://backend.hook.teamofsilicons.com
  description: Backend API; public ingress may use the configured Hook frontend origin.
security:
- bearerAuth: []
...
paths:
...
  "/api/v1/silicons/{silicon_id}/hooks":
    get:
      tags:
      - Hooks
      operationId: listHooks
      summary: List a Silicon's hooks with their URLs and last activity
      parameters:
      - "$ref": "#/components/parameters/OrgId"
      - "$ref": "#/components/parameters/SiliconId"
      - name: include_deleted
        in: query
        schema:
          type: boolean
          default: false
      responses:
        '200':
          description: Active and disabled hooks visible to the actor, plus recoverable
            deleted hooks when include_deleted is true. Each item carries its name,
            endpoint URL, and last_received_at.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HookPage"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '408':
          "$ref": "#/components/responses/RequestTimeout"
        '422':
          "$ref": "#/components/responses/UnprocessableEntity"
        '500':
          "$ref": "#/components/responses/InternalError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
    post:
      tags:
      - Hooks
      operationId: createHook
      summary: Create a signed webhook endpoint
      parameters:
      - "$ref": "#/components/parameters/OrgId"
      - "$ref": "#/components/parameters/SiliconId"
      - "$ref": "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/HookCreate"
      responses:
        '201':
          description: Hook created; the signing secret is shown once.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HookWithSecret"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '408':
          "$ref": "#/components/responses/RequestTimeout"
        '409':
          "$ref": "#/components/responses/Conflict"
        '410':
          "$ref": "#/components/responses/Gone"
        '413':
          "$ref": "#/components/responses/PayloadTooLarge"
        '415':
          "$ref": "#/components/responses/UnsupportedMediaType"
        '422':
          "$ref": "#/components/responses/UnprocessableEntity"
        '500':
          "$ref": "#/components/responses/InternalError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
    patch:
      tags:
      - Hooks
      operationId: setHooksEnabled
      summary: Enable or disable a set of hooks atomically
      description: Applies the desired enabled state to 1–1000 unique retained hooks.
        The operation is atomic and intrinsically idempotent; returned items follow
        the request hook_ids order.
      parameters:
      - "$ref": "#/components/parameters/OrgId"
      - "$ref": "#/components/parameters/SiliconId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/HooksEnabledUpdate"
      responses:
        '200':
          description: All requested hooks now have the requested enabled state.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HookPage"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '408':
          "$ref": "#/components/responses/RequestTimeout"
        '409':
          "$ref": "#/components/responses/Conflict"
        '413':
          "$ref": "#/components/responses/PayloadTooLarge"
        '415':
          "$ref": "#/components/responses/UnsupportedMediaType"
        '422':
          "$ref": "#/components/responses/UnprocessableEntity"
        '500':
          "$ref": "#/components/responses/InternalError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
    parameters:
    - "$ref": "#/components/parameters/TestKey"
    - "$ref": "#/components/parameters/TestAppSecret"
  "/api/v1/silicons/{silicon_id}/hooks/{hook_id}":
    parameters:
    - "$ref": "#/components/parameters/OrgId"
    - "$ref": "#/components/parameters/SiliconId"
    - "$ref": "#/components/parameters/HookId"
    - "$ref": "#/components/parameters/TestKey"
    - "$ref": "#/components/parameters/TestAppSecret"
    get:
      tags:
      - Hooks
      operationId: getHook
      summary: Read one hook
      responses:
        '200':
          description: Hook metadata and signature policy without secret material.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Hook"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '408':
          "$ref": "#/components/responses/RequestTimeout"
        '422':
          "$ref": "#/components/responses/UnprocessableEntity"
        '500':
          "$ref": "#/components/responses/InternalError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
    patch:
      tags:
      - Hooks
      operationId: updateHook
      summary: Change name, description, time zone, activation, or signature policy
      description: 'Every member is optional. `enabled` applies the dedicated activation
        action; any other member is an update. Signature members merge onto the current
        policy; `"public_key": null` clears the key. Supplying `secret` replaces the
        stored secret. Requiring signatures on a symmetric algorithm needs a stored
        or supplied secret.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/HookUpdate"
      responses:
        '200':
          description: Hook after the change.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Hook"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '408':
          "$ref": "#/components/responses/RequestTimeout"
        '409':
          "$ref": "#/components/responses/Conflict"
        '413':
          "$ref": "#/components/responses/PayloadTooLarge"
        '415':
          "$ref": "#/components/responses/UnsupportedMediaType"
        '422':
          "$ref": "#/components/responses/UnprocessableEntity"
        '500':
          "$ref": "#/components/responses/InternalError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
    delete:
      tags:
      - Hooks
      operationId: deleteHook
      summary: Soft-delete a hook
      responses:
        '204':
          description: Hook soft-deleted and retained for 45 days.
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '408':
          "$ref": "#/components/responses/RequestTimeout"
        '413':
          "$ref": "#/components/responses/PayloadTooLarge"
        '422':
          "$ref": "#/components/responses/UnprocessableEntity"
        '500':
          "$ref": "#/components/responses/InternalError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
  "/api/v1/silicons/{silicon_id}/hooks/{hook_id}/restore":
    post:
      tags:
      - Hooks
      operationId: restoreHook
      summary: Restore a recoverable hook
      parameters:
      - "$ref": "#/components/parameters/OrgId"
      - "$ref": "#/components/parameters/SiliconId"
      - "$ref": "#/components/parameters/HookId"
      - "$ref": "#/components/parameters/IdempotencyKey"
      responses:
        '200':
          description: Hook restored with its existing endpoint and secret.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Hook"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '408':
          "$ref": "#/components/responses/RequestTimeout"
        '409':
          "$ref": "#/components/responses/Conflict"
        '410':
          "$ref": "#/components/responses/Gone"
        '413':
          "$ref": "#/components/responses/PayloadTooLarge"
        '422':
          "$ref": "#/components/responses/UnprocessableEntity"
        '500':
          "$ref": "#/components/responses/InternalError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
    parameters:
    - "$ref": "#/components/parameters/TestKey"
    - "$ref": "#/components/parameters/TestAppSecret"
  "/api/v1/silicons/{silicon_id}/hooks/{hook_id}/secret/rotate":
    post:
      tags:
      - Hooks
      operationId: rotateHookSecret
      summary: Issue a new generated signing secret
      parameters:
      - "$ref": "#/components/parameters/OrgId"
      - "$ref": "#/components/parameters/SiliconId"
      - "$ref": "#/components/parameters/HookId"
      - "$ref": "#/components/parameters/IdempotencyKey"
      responses:
        '200':
          description: New secret shown once; the old secret is invalid immediately.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/SigningSecretResponse"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '408':
          "$ref": "#/components/responses/RequestTimeout"
        '409':
          "$ref": "#/components/responses/Conflict"
        '410':
          "$ref": "#/components/responses/Gone"
        '413':
          "$ref": "#/components/responses/PayloadTooLarge"
        '422':
          "$ref": "#/components/responses/UnprocessableEntity"
        '500':
          "$ref": "#/components/responses/InternalError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
    parameters:
    - "$ref": "#/components/parameters/TestKey"
    - "$ref": "#/components/parameters/TestAppSecret"
  "/api/v1/silicons/{silicon_id}/hooks/{hook_id}/endpoint/rotate":
    post:
      tags:
      - Hooks
      operationId: rotateHookEndpoint
      summary: Replace the endpoint key and retire the old one forever
      description: 'Allocates a new eight-character alphanumeric key that has never
        been used for this Silicon. The previous key is retired permanently: it is
        never reissued for the Silicon and requests to it receive 410 endpoint_retired.'
      parameters:
      - "$ref": "#/components/parameters/OrgId"
      - "$ref": "#/components/parameters/SiliconId"
      - "$ref": "#/components/parameters/HookId"
      - "$ref": "#/components/parameters/IdempotencyKey"
      responses:
        '200':
          description: Hook with its new endpoint URL and key.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Hook"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '408':
          "$ref": "#/components/responses/RequestTimeout"
        '409':
          "$ref": "#/components/responses/Conflict"
        '413':
          "$ref": "#/components/responses/PayloadTooLarge"
        '422':
          "$ref": "#/components/responses/UnprocessableEntity"
        '500':
          "$ref": "#/components/responses/InternalError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
    parameters:
    - "$ref": "#/components/parameters/TestKey"
    - "$ref": "#/components/parameters/TestAppSecret"
  "/api/v1/silicons/{silicon_id}/hooks/{hook_id}/events":
    get:
      tags:
      - History
      operationId: listHookEvents
      summary: List the last n verified requests for one hook
      parameters:
      - "$ref": "#/components/parameters/OrgId"
      - "$ref": "#/components/parameters/SiliconId"
      - "$ref": "#/components/parameters/HookId"
      - "$ref": "#/components/parameters/HistoryLimit"
      - "$ref": "#/components/parameters/HistoryCursor"
      responses:
        '200':
          "$ref": "#/components/responses/EventPageOk"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '408':
          "$ref": "#/components/responses/RequestTimeout"
        '422':
          "$ref": "#/components/responses/UnprocessableEntity"
        '500':
          "$ref": "#/components/responses/InternalError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
    parameters:
    - "$ref": "#/components/parameters/TestKey"
    - "$ref": "#/components/parameters/TestAppSecret"
...
  "/api/v1/auth/logout":
    post:
      tags:
      - Auth
      operationId: logout
      summary: End the IAM session behind a Hook Application token
      responses:
        '204':
          description: Session ended.
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '408':
          "$ref": "#/components/responses/RequestTimeout"
        '413':
          "$ref": "#/components/responses/PayloadTooLarge"
        '500':
          "$ref": "#/components/responses/InternalError"
        '503':
          "$ref": "#/components/responses/ServiceUnavailable"
      parameters:
      - "$ref": "#/components/parameters/IdempotencyKey"
      description: Supply a refresh token as bearer to revoke its whole family, or
        an access token to revoke that token. The body is empty.
    parameters:
    - "$ref": "#/components/parameters/TestKey"
    - "$ref": "#/components/parameters/TestAppSecret"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: IAM opaque access token
  parameters:
...
    OrgId:
      name: X-Org-ID
      in: header
      required: true
      schema:
        "$ref": "#/components/schemas/OrganizationId"
    SiliconId:
      name: silicon_id
      in: path
      required: true
      schema:
        "$ref": "#/components/schemas/SiliconId"
...